Update Pseudocode2.cpp
Just realized we can make ReadIn() to be a static member function of the class Network. It can be called directly in the main(), before even declaring any object of the class because it is a static function. And, we can read in the required parameters and assign the values before calling the constructor. So, the constructor doesn't need any parameter if we make ReadIn() a static member function of the class Network.