Added update()
Ended up having to write a mini-constructor to test `update()`, so when it comes down to writing it, I'll take one of the constructors! Also, perhaps we could rename `update()` to something more illustrative? I'm trying to put what Slater said into practice. So, maybe `updateWeightsAndBiases()` or `updateNetwork()`. Considering that we may want to update other parts of the code, it behooves us to change the name.