Update Network.cpp
Added validateInputs(), which checks the values of all the hyperparameters entered by the user or read from the file. It returns a vector of string which contains all the wrong inputs entered by the user or read from the file. It can be used to let user know what went wrong. I have included the prototype of the function in class as a comment for now but we can add it to the class later.