Master's Network.h with a succesful build!
It is running successfully and for now, I have commented out the stuff related to the state table. However the state table is working but it is asking user at every minibatch to input the activation function. So, working on having a function which sets up the activation function for the following minibatches in the epochs.