This won't build until you change certain filepaths and hardcoded hyperparameters, commented with the triple slash (///), so they should be darker. Meaningful Changes made to Source.cpp, activationFuncSelect() and makeGraphFile() in Network.cpp. Some static members added to Network.h to aid in the activation function selection.
Most of this won't be useful, but please take a look at my edits to makeGraphFile(), because I really liked them.

All of the resulting graphs/efficiencies have been added to the GraphFiles directory. The graphs will require that you have graphviz-2.38 (or latest) on your hard drive. The efficiencies should be accessible to any cell-sheet programs such as Google Sheets and Excel.